Move getPreferredAbi method into CuberiteHelper
